# apps/studio-web/openapi/studio-api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "StudioApi | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/events": {
      "get": {
        "tags": [
          "Events"
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "occurrenceId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "publication",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventListItem"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Events"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDetail"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDetail"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftValidationProblem"
                }
              }
            }
          }
        }
      }
    },
    "/api/events/{id}": {
      "get": {
        "tags": [
          "Events"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Events"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftValidationProblem"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventConflict"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminRef": {
        "required": [
          "id",
          "displayName"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          }
        }
      },
...
      "CreateEventRequest": {
        "required": [
          "operationId",
          "draft"
        ],
        "type": "object",
        "properties": {
          "operationId": {
            "type": "string",
            "format": "uuid"
          },
          "draft": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EventDraft"
              }
            ]
          }
        }
      },
...
      "DraftProblem": {
        "required": [
          "path",
          "code",
          "message"
        ],
        "type": "object",
        "properties": {
          "path": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "DraftValidationProblem": {
        "required": [
          "code",
          "title",
          "problems"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "problems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DraftProblem"
            }
          }
        }
      },
...
      "EventConflict": {
        "required": [
          "code",
          "title",
          "latest"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "latest": {
            "$ref": "#/components/schemas/EventDetail"
          }
        }
      },
      "EventDetail": {
        "required": [
          "id",
          "rowVersion",
          "revisionId",
          "createdAt",
          "updatedAt",
          "updatedBy",
          "publication",
          "draft"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "rowVersion": {
            "type": "integer",
            "format": "int64"
          },
          "revisionId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedBy": {
            "$ref": "#/components/schemas/AdminRef"
          },
          "publication": {
            "$ref": "#/components/schemas/PublicationSummary"
          },
          "draft": {
            "$ref": "#/components/schemas/EventDraft"
          }
        }
      },
      "EventDraft": {
        "required": [
          "schemaVersion",
          "title",
          "description",
          "categoryId",
          "occurrenceId",
          "slots"
        ],
        "type": "object",
        "properties": {
          "schemaVersion": {
            "type": "string"
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "categoryId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "occurrenceId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "slots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventSlot"
            }
          }
        }
      },
      "EventListItem": {
        "required": [
          "id",
          "title",
          "occurrenceId",
          "slotCount",
          "venueCount",
          "slots",
          "publication",
          "updatedAt",
          "updatedBy"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "occurrenceId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "slotCount": {
            "type": "integer",
            "format": "int32"
          },
          "venueCount": {
            "type": "integer",
            "format": "int32"
          },
          "slots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventSlotSummary"
            }
          },
          "publication": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedBy": {
            "$ref": "#/components/schemas/AdminRef"
          }
        }
      },
...
      "EventSlot": {
        "required": [
          "slotId",
          "ocDayId",
          "timeMode",
          "fixed",
          "participation",
          "status",
          "cancelNote",
          "venues"
        ],
        "type": "object",
        "properties": {
          "slotId": {
            "type": "string",
            "format": "uuid"
          },
          "ocDayId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "timeMode": {
            "type": [
              "null",
              "string"
            ]
          },
          "fixed": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/FixedTime"
              }
            ]
          },
          "participation": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": "string"
          },
          "cancelNote": {
            "type": [
              "null",
              "string"
            ]
          },
          "venues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SlotVenue"
            }
          }
        }
      },
      "EventSlotSummary": {
        "required": [
          "date",
          "timeMode",
          "start",
          "end",
          "status",
          "dayCancelled"
        ],
        "type": "object",
        "properties": {
          "date": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "timeMode": {
            "type": [
              "null",
              "string"
            ]
          },
          "start": {
            "type": [
              "null",
              "string"
            ]
          },
          "end": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": "string"
          },
          "dayCancelled": {
            "type": "boolean"
          }
        }
      },
      "FixedTime": {
        "required": [
          "start",
          "end"
        ],
        "type": "object",
        "properties": {
          "start": {
            "type": [
              "null",
              "string"
            ]
          },
          "end": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "PublicationSummary": {
        "required": [
          "state",
          "publishedRevisionId",
          "publishedAt"
        ],
        "type": "object",
        "properties": {
          "state": {
            "type": "string"
          },
          "publishedRevisionId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "publishedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "SlotVenue": {
        "required": [
          "canonicalSpotId",
          "note"
        ],
        "type": "object",
        "properties": {
          "canonicalSpotId": {
            "type": "string"
          },
          "note": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "UpdateEventRequest": {
        "required": [
          "operationId",
          "rowVersion",
          "draft"
        ],
        "type": "object",
        "properties": {
          "operationId": {
            "type": "string",
            "format": "uuid"
          },
          "rowVersion": {
            "type": "integer",
            "format": "int64"
          },
          "draft": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EventDraft"
              }
            ]
          }
        }
      },
...
    }
  },
...
}